requirements.txt
openai>=1.45.0
gradio>=4.44.0
pillow>=10.4.0
requests>=2.32.3